requirements.txt
# 核心依赖
playwright>=1.40.0
beautifulsoup4>=4.0.0
requests>=2.32.0
Pillow>=10.0.0
qrcode>=7.0.0

# 可选依赖
lxml>=4.0.0
